Adding test + renaming method
